Create test_lcc_classifier..json
These are test cases.  Fix any that are wrong and add any new one that you can think of.

We put the LCC code into our translator then it should return the corresponding list of classifiers.

The more different test cases that we have in this list, the more we can find and fix flaws in our translator.  https://github.com/cclauss/library-of-congress-classifier-hack